カード型に整形
