+++ components/Header.js
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <header className="w-screen h-20 bg-blue-100 pl-5 shadow-sm">
+      <h1 className="font-bold text-3xl leading-[80px]">React TODO</h1>
+    </header>
+  );
+};
+
+export default Header;
